next

+++ src/05-Data_Collection.cuh
@@ -6,9 +6,9 @@
  */
 typedef struct SmallDataCollection {
     // small = 16 bit
-    float mulp[65536];
-    float warp[65536];
-    float lane[65536];
+    int mulp[65536];
+    int warp[65536];
+    int lane[65536];
     float time[65536];
 } SmallDataCollection;
 
@@ -18,9 +18,9 @@ typedef struct SmallDataCollection {
  */
 typedef struct MediumDataCollection {
     // medium = 24 bit
-    float mulp[16777216];
-    float warp[16777216];
-    float lane[16777216];
+    int mulp[16777216];
+    int warp[16777216];
+    int lane[16777216];
     float time[16777216];
 } MediumDataCollection;
 
@@ -30,9 +30,9 @@ typedef struct MediumDataCollection {
  */
 typedef struct LargeDataCollection {
     // large = 32 bit
-    float mulp[4294967296];
-    float warp[4294967296];
-    float lane[4294967296];
+    int mulp[4294967296];
+    int warp[4294967296];
+    int lane[4294967296];
     float time[4294967296];
 } LargeDataCollection;
